+++ test/MicroElements.Metadata.Tests/XmlReader/XmlReaderTests.cs
@@ -248,7 +248,7 @@ namespace MicroElements.Metadata.Tests
             List<IPropertyContainer> list = new List<IPropertyContainer>();
 
             var schema = new PropertySet().ToSchema();
-            IXmlParserSettings parserSettings = new XmlParserSettings();
+            IXmlParserSettings parserSettings = new XmlParserSettings(propertyValueFactory: new PropertyValueFactory());
             IXmlParserContext parserContext = new XmlParserContext(parserSettings, schema);
 
             foreach (string file in Directory