+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+# 7.0.0-rc.22
+- Performance optimization for metadata search
+- IPropertyValueFactory, CachedPropertyValueFactory
+- XmlParser: IPropertyValueFactory added to IXmlParserSettings
+- CachedStringParser and InernedStringParser
+
 # 7.0.0-rc.19
 - Validation for XmlParser
 - ConfigureMetadata now supports thread safe ReadOnly metadata creation and configuration